refactor : 경로 설정
